Working on the board logic. Monte Carlo basics should be simple to implement, along with a policy, after  the move generation is up functional

+++ Gopher/Random.h
@@ -3,8 +3,8 @@
 
 static unsigned long RSeed = 29264;
 
-uint16_t fastRandom(unsigned int max)
+int fastRandom(unsigned int max)
 {
 	RSeed = ((RSeed * 1103515245) + 12345) & 0x7fffffff;
-	return ((RSeed & 0xffff) * max) >> 16;
+	return static_cast<int>(((RSeed & 0xffff) * max) >> 16);
 }
